requirements.txt
allure-pytest~=2.13.5
allure-python-commons~=2.13.5
Appium-Python-Client~=4.4.0
pydantic-settings==2.6.1
pytest~=8.3.4
selene~=2.0.0rc9
selenium~=4.27.1